Upgrade ppdiffusers0240's models slow tests (#396)

+++ ppdiffusers/ppdiffusers/models/consistency_decoder_vae.py
@@ -329,7 +329,7 @@ class ConsistencyDecoderVAE(ModelMixin, ConfigMixin):
         )
 
         for t in self.decoder_scheduler.timesteps:
-            model_input = paddle.concat([self.decoder_scheduler.scale_model_input(x_t, t), z], axis=1)
+            model_input = paddle.concat([self.decoder_scheduler.scale_model_input(x_t, t).cast(z.dtype), z], axis=1)
             model_output = self.decoder_unet(model_input, t).sample[:, :3, :, :]
             prev_sample = self.decoder_scheduler.step(model_output, t, x_t, generator).prev_sample
             x_t = prev_sample
